+++ app/adapters/avito.py
@@ -1,8 +1,8 @@
 import httpx
 from typing import Optional
 
-from app.config import settings
-from app.oauth2 import ensure_fresh_access
+from app.core.config import settings
+from app.api.oauth2 import ensure_fresh_access
 from app.core.retry import with_retry
